@@ -2,7 +2,7 @@ package cn.imrhj.sharetoqrcode.ui.activity
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 
 class LoadingActivity : AppCompatActivity() {
 
