@@ -4,7 +4,7 @@ import android.content.Context
 import android.content.Intent
 import android.media.projection.MediaProjectionManager
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.util.Log
 import cn.imrhj.sharetoqrcode.service.CaptureScreenService
 import cn.imrhj.sharetoqrcode.util.transparentStatusBar
